Add files via upload
Handling dietary preferences in a new culture can be challenging due to unfamiliar language and indecipherable ingredients. This project aims to solve this issue faced by individuals moving globally. We have developed a classification system into three fundamental categories: Vegetarian, Non-vegetarian, Eggiterian. Leveraging the machine learning algorithms and optical character recognition libraries that already exist, our model gives a high accuracy in classifying a list of ingredients into the three classes. This could help an individual adapt in various countries without having to worry about his/her dietary preferences and having to learn another language. The classification of the ingredients list helps individuals, especially those who travel internationally very often to make informed decisions about their diet according to their cultural and ethical beliefs. It eases the load of having to manually type in the list of ingredients on the food packet. Our project extracts the list of ingredients from the image provided and performs the classification. The classifier is trained on a variety of packaged food and diets from all around the world which makes it robust and a trustable machine and lays the groundwork for further extension. In our project, we aim to address a common problem faced by many individuals with a practical solution that helps them choose and adapt to different foods of different countries.